Adding useStepper hook to handle Stepper with Jotai
