+++ src/atoms/current-step-atom.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai';
+
+const currentStepAtom = atom(1);
+
+export { currentStepAtom };
